ver1.0のアップ

+++ towerDiffence/TowerDiffence/towerDiffence.cpp
@@ -944,7 +944,7 @@ void TowerDiffence::loadHighScore()
 		{
 			if (getline(stream, token, ','))
 			{
-				mHighScores[i] = stof(token);
+				mHighScores[i] = (int)stof(token);
 			}
 		}
 	}
